backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Cache
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML
# openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
APScheduler==3.10.4
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Deployment
gunicorn==21.2.0
docker==6.1.3